refactor: move processing packages to internal
