+++ internal/align/fuzz_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclwrite"
-	"github.com/oferchen/hclalign/hclprocessing"
+	"github.com/oferchen/hclalign/internal/hclalign"
 )
 
 func FuzzReorderStability(f *testing.F) {
@@ -37,7 +37,7 @@ func FuzzReorderStability(f *testing.F) {
 		if diags.HasErrors() {
 			t.Fatalf("parse: %v", diags)
 		}
-		hclprocessing.ReorderAttributes(file, nil, false)
+		hclalign.ReorderAttributes(file, nil, false)
 		out := file.Bytes()
 
 		if len(out) > maxFuzzBytes {
@@ -48,7 +48,7 @@ func FuzzReorderStability(f *testing.F) {
 		if diags.HasErrors() {
 			t.Fatalf("parse reordered: %v", diags)
 		}
-		hclprocessing.ReorderAttributes(file2, nil, false)
+		hclalign.ReorderAttributes(file2, nil, false)
 		if !bytes.Equal(out, file2.Bytes()) {
 			t.Fatalf("round-trip mismatch")
 		}
