+++ internal/engine/engine_test.go
@@ -1,5 +1,5 @@
-// fileprocessing/fileprocessing_test.go
-package fileprocessing
+// internal/engine/engine_test.go
+package engine
 
 import (
 	"bytes"
@@ -14,8 +14,8 @@ import (
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclwrite"
 	"github.com/oferchen/hclalign/config"
-	"github.com/oferchen/hclalign/hclprocessing"
 	internalfs "github.com/oferchen/hclalign/internal/fs"
+	"github.com/oferchen/hclalign/internal/hclalign"
 )
 
 func TestProcessPreservesNewlineAndBOM(t *testing.T) {
@@ -324,7 +324,7 @@ func TestProcessReaderPreservesNewlineAndBOM(t *testing.T) {
 	if diags.HasErrors() {
 		t.Fatalf("parse expected: %v", diags)
 	}
-	hclprocessing.ReorderAttributes(expectedFile, config.CanonicalOrder, false)
+	hclalign.ReorderAttributes(expectedFile, config.CanonicalOrder, false)
 	expected := internalfs.ApplyHints(expectedFile.Bytes(), internalfs.Hints{HasBOM: true, Newline: "\r\n"})
 	if string(out) != string(expected) {
 		t.Fatalf("unexpected output: got %q, want %q", out, expected)
@@ -413,8 +413,6 @@ func TestProcessStopsAfterMalformedFile(t *testing.T) {
 	if err != nil {
 		t.Fatalf("read good file: %v", err)
 	}
-	// The good file should remain unmodified since processing stops after
-	// encountering the malformed file and the write operation is canceled.
 	expected := []byte("variable \"a\" {\n  default = 1\n  type = number\n}\n")
 	if string(data) != string(expected) {
 		t.Fatalf("good file unexpectedly processed: got %q, want %q", data, expected)
