+++ internal/engine/testdata/idempotent_golden.hcl
