+++ internal/engine/testdata/idempotent_input.hcl
